requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
config==0.5.1
contourpy==1.2.0; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
dateparser==1.1.8; python_version >= '3.7'
flask==3.0.0; python_version >= '3.8'
fonttools==4.44.0; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.8.1; python_version >= '3.9'
multidict==6.0.4; python_version >= '3.7'
numpy==1.26.1; python_version < '3.13' and python_version >= '3.9'
packaging==23.2; python_version >= '3.7'
pandas==2.1.2; python_version >= '3.9'
pillow==10.1.0; python_version >= '3.8'
plotly==5.18.0; python_version >= '3.6'
pycryptodome==3.19.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-binance==1.0.19
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ta-lib==0.4.28
tenacity==8.2.3; python_version >= '3.7'
tzdata==2023.3; python_version >= '2'
tzlocal==5.2; python_version >= '3.8'
ujson==5.8.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
websockets==12.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
yarl==1.9.2; python_version >= '3.7'
